+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## unreleased
 
+## 0.6.0
+
 * `tada help`: commands in the alphabetical order
 * bring back `tada cache`: install package to the local cache